+++ webargs_starlette/starletteparser.py
@@ -11,7 +11,7 @@ from webargs.asyncparser import AsyncParser
 from webargs import core
 from .annotations import TypeMapping, DEFAULT_TYPE_MAPPING, annotations2schema
 
-HTTP_METHOD_NAMES = [
+HTTP_METHOD_NAMES: typing.List[str] = [
     "get",
     "post",
     "put",
@@ -54,7 +54,7 @@ class StarletteParser(AsyncParser):
 
     TYPE_MAPPING: TypeMapping = DEFAULT_TYPE_MAPPING
 
-    __location_map__: typing.Mapping[str, str] = dict(
+    __location_map__: typing.Dict[str, typing.Union[str, typing.Callable]] = dict(
         path_params="parse_path_params", **core.Parser.__location_map__
     )
 
@@ -90,7 +90,7 @@ class StarletteParser(AsyncParser):
             self._cache["json"] = json_data
         return core.get_value(json_data, name, field, allow_many_nested=True)
 
-    async def parse_form(self, req, name, field) -> typing.Any:
+    async def parse_form(self, req: Request, name: str, field: Field) -> typing.Any:
         """Pull a form value from the request."""
         post_data = self._cache.get("form")
         if post_data is None:
@@ -104,7 +104,9 @@ class StarletteParser(AsyncParser):
             400, exception=error, messages={"json": ["Invalid JSON body."]}
         )
 
-    def get_request_from_view_args(self, view, args, kwargs) -> Request:
+    def get_request_from_view_args(
+        self, view: typing.Callable, args: tuple, kwargs: dict
+    ) -> Request:
         """Get request object from a handler function or method. Used internally by
         ``use_args`` and ``use_kwargs``.
         """
@@ -142,7 +144,7 @@ class StarletteParser(AsyncParser):
         *,
         type_mapping: TypeMapping = None,
         **kwargs,
-    ) -> typing.Union[typing.Callable, HTTPEndpoint]:
+    ) -> typing.Union[typing.Callable[..., typing.Callable], HTTPEndpoint]:
         # Allow using this as either a decorator or a decorator factory.
         if fn is None:
             return functools.partial(
@@ -150,7 +152,7 @@ class StarletteParser(AsyncParser):
             )
         type_mapping = type_mapping or self.TYPE_MAPPING
 
-        def decorator(func: typing.Callable):
+        def decorator(func: typing.Callable) -> typing.Callable:
             schema = annotations2schema(func, type_mapping=type_mapping)
 
             @functools.wraps(func)